- Created PrintFileVisitor. - Is now possible to copy a directory into a file. A tree of each file contained in the srcDir will be printed inside the destFile. - Refactoring

+++ src/main/java/com/matteoveroni/javacopier/JavaCopier.java
@@ -2,8 +2,9 @@ package com.matteoveroni.javacopier;
 
 import com.matteoveroni.javacopier.copyhistory.CopyHistory;
 import com.matteoveroni.javacopier.filevisitors.CopyDirsFileVisitor;
-import com.matteoveroni.javacopier.filevisitors.CountFilesVisitor;
+import com.matteoveroni.javacopier.filevisitors.CountFileVisitor;
 import com.matteoveroni.javacopier.copyhistory.CopyHistoryEvent;
+import com.matteoveroni.javacopier.filevisitors.PrintFileVisitor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,35 +16,39 @@ import java.nio.file.*;
  */
 public class JavaCopier {
 
-    public static final CopyOption[] STANDARD_COPY_OPTIONS = new CopyOption[]{StandardCopyOption.COPY_ATTRIBUTES};
+    static final CopyOption[] DEFAULT_COPY_OPTIONS = new CopyOption[]{StandardCopyOption.COPY_ATTRIBUTES};
     static final String ERROR_MSG_SRC_OR_DEST_NULL = "src and dest cannot be null";
     static final String ERROR_MSG_SRC_MUST_EXIST = "src must exist";
-    static final String ERROR_MSG_CANNOT_COPY_DIR_INTO_FILE = "cannot copy a directory into a file";
     private static final Logger LOG = LoggerFactory.getLogger(JavaCopier.class);
 
-    public static CopyStatusReport copy(Path src, Path dest, CopyOption... copyOptions) throws IllegalArgumentException, IOException {
-        return copy(src, dest, null, copyOptions);
+    public static CopyStatusReport copy(Path src, Path dest, CopyOption... copyOptions) throws IllegalArgumentException {
+        return copy(src, dest, null, null, copyOptions);
     }
 
-    public static CopyStatusReport copy(Path src, Path dest, CopyListener copyListener, CopyOption... copyOptions) throws IllegalArgumentException, IOException {
+    public static CopyStatusReport copy(Path src, Path dest, CopyListener copyListener, CopyOption... copyOptions) throws IllegalArgumentException {
+        return copy(src, dest, copyListener, null, copyOptions);
+    }
+
+    public static CopyStatusReport copy(Path src, Path dest, OutputStream logReportOutputStream, CopyOption... copyOptions) throws IOException {
+        return copy(src, dest, null, logReportOutputStream, copyOptions);
+    }
+
+    public static CopyStatusReport copy(Path src, Path dest, CopyListener copyListener, OutputStream logReportOutputStream, CopyOption... copyOptions) throws IllegalArgumentException {
         if (src == null || dest == null) {
             throw new IllegalArgumentException(ERROR_MSG_SRC_OR_DEST_NULL);
         } else if (Files.notExists(src)) {
             throw new IllegalArgumentException(ERROR_MSG_SRC_MUST_EXIST);
-        } else if (src.toFile().isDirectory() && (Files.exists(dest) && dest.toFile().isFile())) {
-            throw new IllegalArgumentException(ERROR_MSG_CANNOT_COPY_DIR_INTO_FILE);
         }
 
         LOG.debug("calculating the number of files to copy...");
         Integer totalFiles = calculateFilesCount(src);
         LOG.debug("number of files to copy: " + totalFiles);
 
-        copyOptions = (copyOptions != null) ? copyOptions : STANDARD_COPY_OPTIONS;
-
-        return executeCopy(src, dest, copyListener, totalFiles, copyOptions);
+        copyOptions = (copyOptions != null) ? copyOptions : DEFAULT_COPY_OPTIONS;
+        return executeCopy(src, dest, copyListener, totalFiles, logReportOutputStream, copyOptions);
     }
 
-    private static CopyStatusReport executeCopy(Path src, Path dest, CopyListener copyListener, Integer totalFiles, CopyOption[] copyOptions) {
+    private static CopyStatusReport executeCopy(Path src, Path dest, CopyListener copyListener, Integer totalFiles, OutputStream logReportOutputStream, CopyOption[] copyOptions) {
         LOG.debug("Copy from src: " + src + " to dest: " + dest + " started");
         CopyHistory copyHistory = new CopyHistory();
         CopyStatusReport copyStatus;
@@ -51,12 +56,23 @@ public class JavaCopier {
         try {
             if (src.toFile().isFile() && (Files.notExists(dest) || dest.toFile().isFile())) {
                 Files.copy(src, dest, copyOptions);
-            } else if (src.toFile().isFile() && dest.toFile().isDirectory()) {
+            } else if (src.toFile().isFile() && (Files.exists(dest) && dest.toFile().isDirectory())) {
                 Files.copy(src, Paths.get(dest + File.separator + src.toFile().getName()), copyOptions);
             } else if (src.toFile().isDirectory() && (Files.notExists(dest) || dest.toFile().isDirectory())) {
                 isCopyMultiple = true;
                 CopyDirsFileVisitor copyDirsFileVisitor = new CopyDirsFileVisitor(src, dest, totalFiles, copyHistory, copyListener, copyOptions);
                 Files.walkFileTree(src, copyDirsFileVisitor);
+            } else if (src.toFile().isDirectory() && (Files.exists(dest) && dest.toFile().isFile())) {
+                isCopyMultiple = true;
+                PrintFileVisitor printFileVisitor = null;
+                try {
+                    printFileVisitor = new PrintFileVisitor(src, dest, new FileOutputStream(dest.toFile()), totalFiles, copyHistory, copyListener, copyOptions);
+                    Files.walkFileTree(src, printFileVisitor);
+                } finally {
+                    if (printFileVisitor != null) {
+                        printFileVisitor.closeOutputStream();
+                    }
+                }
             }
             if (!isCopyMultiple) {
                 registerSingleCopySuccessToHistory(src, dest, copyHistory);
@@ -69,20 +85,21 @@ public class JavaCopier {
         }
         copyStatus = new CopyStatusReport(src, dest, CopyStatusReport.CopyState.DONE, totalFiles, copyHistory, copyOptions);
         notifyCopyStatusToListener(copyStatus, copyListener);
-
-        // TODO: log copyStatus to a file. Now the code is hardcoded. Pass an outputstream to copy from outside
-        try (PrintWriter p = new PrintWriter(new FileOutputStream("/home/mavek/java-copier-log.txt", true))) {
-            p.println(copyStatus.toPrettyString());
-        } catch (FileNotFoundException ex) {
-            ex.printStackTrace();
-        }
+        logCopyReportStatusToOutputStream(logReportOutputStream, copyStatus);
         return copyStatus;
     }
 
-    private static Integer calculateFilesCount(Path src) throws IOException {
-        CountFilesVisitor fileCounter = new CountFilesVisitor();
-        Files.walkFileTree(src, fileCounter);
-        return fileCounter.getFilesCount();
+    private static Integer calculateFilesCount(Path src) {
+        int filesCount;
+        try {
+            CountFileVisitor fileCounter = new CountFileVisitor();
+            Files.walkFileTree(src, fileCounter);
+            filesCount = fileCounter.getFilesCount();
+        } catch (IOException ex) {
+            LOG.debug("Error during files count. This should not happen because fileCounterVisitor doesnt throw ioexceptions. ex: " + ex);
+            filesCount = 0;
+        }
+        return filesCount;
     }
 
     private static void registerSingleCopySuccessToHistory(Path src, Path dest, CopyHistory copyHistory) {
@@ -104,6 +121,16 @@ public class JavaCopier {
     private static void notifyCopyStatusToListener(CopyStatusReport copyStatus, CopyListener copyListener) {
         if (copyListener != null) {
             copyListener.onCopyComplete(copyStatus);
+        }
+    }
+
+    private static void logCopyReportStatusToOutputStream(OutputStream logOutputStream, CopyStatusReport copyStatus) {
+        if (logOutputStream != null) {
+            try (PrintWriter printWriter = new PrintWriter(logOutputStream, true)) {
+                printWriter.println(copyStatus.toPrettyString());
+            } catch (Exception ex) {
+                LOG.error("Error trying to log copy status report to outputstream. ex: ", ex);
+            }
         }
     }
 }
